requirements.txt
pydirectinput==1.0.4
Pillow==10.0.0
pyautogui==0.9.54
